refactor(-): :hammer: move testes logs dietas
