+++ sme_sigpae_api/dieta_especial/__tests__/test_serializers.py
@@ -46,21 +46,6 @@ def test_solicitacao_recrei_nas_ferias_serializer(
     assert "dieta_alterada" in data
 
 
-def test_log_dietas_recreio_nas_ferias_serializer(
-    logs_dieta_recreio_nas_ferias,
-):
-    log_recreio = LogQuantidadeDietasAutorizadasRecreioNasFerias.objects.first()
-    log = LogQuantidadeDietasAutorizadasRecreioNasFeriasSerializer(log_recreio)
-    data = log.data
-    assert data is not None
-    assert "escola" in data
-    assert "classificacao" in data
-    assert "dia" in data
-    assert "criado_em" in data
-    assert "data" in data
-    assert "quantidade" in data
-
-
 def test_serializer_solicitacao_dieta_especial_relatorio_terceirizada_fallback(
     solicitacao_dieta_especial,
     protocolo_padrao_dieta_especial,
